requirements.txt
# Python requirements for Trump Podcast Generator TTS
# Install with: pip install -r requirements.txt

# Core TTS Engine
tortoise-tts>=2.4.0

# Audio Processing
torch>=1.13.0
torchaudio>=0.13.0
numpy>=1.21.0
scipy>=1.7.0

# GPU Acceleration (optional but recommended)
# Uncomment if you have CUDA GPU:
# torch>=1.13.0+cu117 -f https://download.pytorch.org/whl/torch_stable.html
# torchaudio>=0.13.0+cu117 -f https://download.pytorch.org/whl/torch_stable.html

# Audio Format Support
librosa>=0.9.0
soundfile>=0.10.0

# Utilities
transformers>=4.20.0
tokenizers>=0.12.0

# Optional: Voice Cloning Enhancement
# unidecode>=1.3.0
# inflect>=5.6.0

# Development Tools (optional)
# jupyter>=1.0.0
# matplotlib>=3.5.0
